MOD: Move marker visulization related methods to MarkerViewer
